blog/requirements.txt
Django==2.2.13
django-cors-headers==3.8.0
djangorestframework==3.12.4
pytz==2018.5
sqlparse==0.4.2
